Split PublicationSanityCheck
